Modified configuration to allow arbitrary code
This makes it easier to configure which urls are going to get hit from
the workers. The person configuring the master and managers can setup a
function to return exactly the urls they want to have requested in the
exact order they need, without having to store them all in an array. The
problem is that they can no longer set the configuration on the master
and have it propagate to the managers, since this would require the
transfer of code across the network and then using something like eval
or pickle to restore it, which is insecure.